requirements.txt
Flask
gunicorn
Flask-SQLAlchemy
Flask-Migrate
itsdangerous
MarkupSafe
Werkzeug
psycopg2
Flask-Script
click
flask_apscheduler
threading
flask_mail
